replacing {{...}} with eval expression, instead of eval the whole response
