+++ lib/api/interceptor.js
@@ -6,7 +6,7 @@ import MockedRequests from 'api/mocked-requests';
 
 import STATUS_CODES from 'api/constants/status-codes';
 import EVENTS from 'api/constants/events';
-import { evalResponse } from 'api/eval-response';
+import { transformResponse } from 'api/utils/transform-response';
 
 class XHRInterceptor {
 
@@ -48,12 +48,12 @@ class XHRInterceptor {
   }
 
   buildResponse(responder, responseRule, xhrRequest) {
-    //let responseBody = evalResponse(responseRule, xhrRequest);
+    let responseBody = transformResponse(responseRule, xhrRequest);
     const response = {
       status: Number(responseRule.status) || 200,
       statusText: STATUS_CODES[responseRule.status || 200].toUpperCase(),
-      data: responseRule.body,
-      text: responseRule.body,
+      data: responseBody,
+      text: responseBody,
       headers: JSON.parse(responseRule.headers || '{}')
     };
